Stop calling error_led.sh when in reboot/shutdown process

+++ varbak.sh
@@ -495,9 +495,13 @@ if [[ $memused -ge $mempeak ]]; then
    # Sync data by using /varbak on CF
    echo "INFO: Using varbak on CF because used RAM greater than configured RAM peak." >>"$lf"
    
-   # Activate warning led because memused is greater than mempeak
-   echo "INFO: Calling $error_led with --warning parameter ..." >>"$lf"
-   "$error_led" --warning "$lf" & 
+   
+   # Activate warning led because memused is greater than mempeak.
+   # But don't do it if we are in reboot/shutdown process.
+   if [[ "$syshalt" != "stop" ]]; then
+      echo "INFO: Calling $error_led with --warning parameter ..." >>"$lf"
+      "$error_led" --warning "$lf" & 
+   fi
 
    # Stop daemons/non-daemons
    pstopper
@@ -515,9 +519,12 @@ elif [[ $memused -lt $mempeak ]] && [[ $varsize -gt $maxtmpfssize ]]; then
      # Use /varbak on CF because we dont want to accupy to much RAM.
      echo "INFO: Using varbak on CF because we do not have enough free RAM." >>"$lf"
     
-     # Activate warning led because varsize is greater than maxtmpfssize
-     echo "INFO: Calling $error_led with --warning parameter ..." >>"$lf"
-     "$error_led" --warning "$lf" & 
+     # Activate warning led because varsize is greater than maxtmpfssize.
+     # But don't do it if we are in reboot/shutdown process.
+     if [[ "$syshalt" != "stop" ]]; then
+         echo "INFO: Calling $error_led with --warning parameter ..." >>"$lf"
+         "$error_led" --warning "$lf" & 
+     fi
 
      # Stop daemons/non-daemons
      pstopper
@@ -534,12 +541,15 @@ else
      echo "INFO: Using varbak on tmpfs because there is enough free RAM." >>"$lf"
 
      # Deactivate warning led because we have enough free RAM.
-     if [[ $(pgrep $(basename "$error_led")) ]]; then 
-        echo "KILLALL: Killing $error_led ..." >>"$lf"
-        killall -e -9 `basename "$error_led"` >>"$lf" 2>&1
-        echo "KILLALL: Done." >>"$lf"
-        echo "INFO: Calling $error_led with --warn-off parameter ..." >>"$lf"
-        "$error_led" --warn-off "$lf" 
+     # But don't do it if we are in reboot/shutdown process.
+     if [[ "$syshalt" != "stop" ]]; then
+        if [[ $(pgrep $(basename "$error_led")) ]]; then 
+           echo "KILLALL: Killing $error_led ..." >>"$lf"
+           killall -e -9 `basename "$error_led"` >>"$lf" 2>&1
+           echo "KILLALL: Done." >>"$lf"
+           echo "INFO: Calling $error_led with --warn-off parameter ..." >>"$lf"
+           "$error_led" --warn-off "$lf" 
+        fi
      fi
      
      # Mount tmpfs for data sync.
